requirements.txt
Babel==2.8.0
click==7.1.2
Flask==1.1.1
Flask-Babel==1.0.0
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.20
protobuf==3.12.1
PyMySQL==0.9.3
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
